fix: split client currency hooks from server-safe module
